requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
greenlet==3.0.1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
